chore: chore is now aware of orderSymbol
